+++ src/routes/board/comment/[commentId]/index.tsx
@@ -1,6 +1,7 @@
 import { component$, useSignal, useTask$ } from "@builder.io/qwik";
 import { Link, routeLoader$, type DocumentHead } from "@builder.io/qwik-city";
 import type { Comment } from "@prisma/client";
+import { trpcFetch$ } from "~/lib/qwik-trpc2";
 import { CommentActions } from "~/modules/comment/CommentActions/CommentActions";
 import { CommentsList } from "~/modules/comment/CommentsList/CommentsList";
 import { CreateCommentForm } from "~/modules/comment/CreateCommentForm/CreateCommentForm";
@@ -19,7 +20,7 @@ export const useComments = routeLoader$((event) =>
   })
 );
 
-// const queryMoreComments = server$(trpc.comment.listForParent.query());
+const queryMoreComments = trpcFetch$(() => ["comment", "listForParent"]);
 
 type CommentCardProps = {
   comment: Comment;
@@ -62,11 +63,12 @@ export const CommentCard = component$<CommentCardProps>((props) => {
       <button
         class="btn"
         onClick$={async () => {
-          // const value = await queryMoreComments({
-          //   parentId: props.comment.id,
-          //   skip: (page.value + 1) * 10,
-          //   take: 10,
-          // });
+          const value = await queryMoreComments({
+            parentId: props.comment.id,
+            skip: (page.value + 1) * 10,
+            take: 10,
+          });
+          console.log({ value });
           // if (value.status === "success") {
           //   const nextCollection = [...collection.value];
           //   nextCollection.push(...value.result.comments);
