+++ src/routes/board/post/[postId]/index.tsx
@@ -1,6 +1,7 @@
 import { component$, useSignal, useTask$ } from "@builder.io/qwik";
 import { Link, routeLoader$, type DocumentHead } from "@builder.io/qwik-city";
 import type { Comment, Post } from "@prisma/client";
+import { trpcFetch$ } from "~/lib/qwik-trpc2";
 import { CommentsList } from "~/modules/comment/CommentsList/CommentsList";
 import { CreateCommentForm } from "~/modules/comment/CreateCommentForm/CreateCommentForm";
 import { PostActions } from "~/modules/post/PostActions/PostActions";
@@ -19,7 +20,7 @@ export const useComments = routeLoader$((event) =>
   })
 );
 
-// const queryMoreComments = server$(trpc.comment.listForPost.query());
+const queryMoreComments = trpcFetch$(() => ["comment", "listForPost"]);
 
 type PostCardProps = {
   post: Post;
@@ -55,11 +56,12 @@ export const PostCard = component$<PostCardProps>((props) => {
       <button
         class="btn"
         onClick$={async () => {
-          // const value = await queryMoreComments({
-          //   postId: props.post.id,
-          //   skip: (page.value + 1) * 10,
-          //   take: 10,
-          // });
+          const value = await queryMoreComments({
+            postId: props.post.id,
+            skip: (page.value + 1) * 10,
+            take: 10,
+          });
+          console.log(value);
           // if (value.status === "success") {
           //   const nextCollection = [...collection.value];
           //   nextCollection.push(...value.result.comments);
